# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Roleplay Coach Hub
  version: 0.1.0
paths:
...
  /api/v1/simulator/start:
    post:
      tags:
      - simulator
      summary: Start Dialogue
      description: Start a new simulator dialogue (standalone).
      operationId: start_dialogue_api_v1_simulator_start_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartDialogueRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartDialogueResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/simulator/respond:
    post:
      tags:
      - simulator
      summary: Generate Response
      description: Generate a client response for a session.
      operationId: generate_response_api_v1_simulator_respond_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RespondRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RespondResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/simulator/should-end/{session_id}:
    post:
      tags:
      - simulator
      summary: Should End
      description: Check if the dialogue should end.
      operationId: should_end_api_v1_simulator_should_end__session_id__post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Session Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: boolean
                title: Response Should End Api V1 Simulator Should End  Session Id  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/curator/learning-plan:
    post:
      tags:
      - curator
      summary: Generate Learning Plan
      description: Generate a learning plan based on evaluation history.
      operationId: generate_learning_plan_api_v1_curator_learning_plan_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LearningPlanRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningPlan'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/curator/quiz:
    post:
      tags:
      - curator
      summary: Generate Quiz
      description: Generate a micro-quiz for a scenario (trainer+).
      operationId: generate_quiz_api_v1_curator_quiz_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuizResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/curator/sync-lms:
    post:
      tags:
      - curator
      summary: Sync To Lms
      description: 'Stub: sync a learning plan to LMS (admin only).'
      operationId: sync_to_lms_api_v1_curator_sync_lms_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncLMSRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Sync To Lms Api V1 Curator Sync Lms Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/xp/{user_id}:
    get:
      tags:
      - gamification
      summary: Get Xp Balance
      description: Return XP balance + level for a user.
      operationId: get_xp_balance_api_v1_gamification_xp__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Xp Balance Api V1 Gamification Xp  User Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/xp/{user_id}/history:
    get:
      tags:
      - gamification
      summary: Get Xp History
      description: Return paginated XP transaction history.
      operationId: get_xp_history_api_v1_gamification_xp__user_id__history_get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_dict_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/badges:
    get:
      tags:
      - gamification
      summary: List All Badges
      description: Return all available badges.
      operationId: list_all_badges_api_v1_gamification_badges_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List All Badges Api V1 Gamification Badges Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/badges/{user_id}:
    get:
      tags:
      - gamification
      summary: Get User Badges
      description: Return badges earned by a user.
      operationId: get_user_badges_api_v1_gamification_badges__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get User Badges Api V1 Gamification Badges  User Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/leaderboard:
    get:
      tags:
      - gamification
      summary: Get Leaderboard
      description: Return top users by XP.
      operationId: get_leaderboard_api_v1_gamification_leaderboard_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Size
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_dict_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/gamification/streak/{user_id}:
    get:
      tags:
      - gamification
      summary: Get Streak
      description: Return current streak for a user.
      operationId: get_streak_api_v1_gamification_streak__user_id__get
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Streak Api V1 Gamification Streak  User Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LearningPlan:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        scenario_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scenario Id
        focus_areas:
          items:
            type: string
          type: array
          title: Focus Areas
        steps:
          items:
            $ref: '#/components/schemas/PlanStep'
          type: array
          title: Steps
        difficulty_label:
          type: string
          title: Difficulty Label
          default: beginner
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - user_id
      title: LearningPlan
      description: Personalised learning plan based on evaluation weaknesses.
    LearningPlanRequest:
      properties:
        scenario_id:
          type: string
          format: uuid
          title: Scenario Id
      type: object
      required:
      - scenario_id
      title: LearningPlanRequest
...
    MicroQuiz:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        scenario_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Scenario Id
        title:
          type: string
          minLength: 1
          title: Title
        questions:
          items:
            $ref: '#/components/schemas/QuizQuestion'
          type: array
          title: Questions
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - title
      title: MicroQuiz
      description: A short quiz tied to a scenario for skill reinforcement.
...
    Page_dict_:
      properties:
        items:
          items: {}
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        size:
          type: integer
          title: Size
      type: object
      required:
      - items
      - total
      - page
      - size
      title: Page[dict]
    PlanStep:
      properties:
        order:
          type: integer
          minimum: 1.0
          title: Order
        title:
          type: string
          minLength: 1
          title: Title
        description:
          type: string
          minLength: 1
          title: Description
        estimated_minutes:
          type: integer
          minimum: 1.0
          title: Estimated Minutes
          default: 10
      type: object
      required:
      - order
      - title
      - description
      title: PlanStep
      description: A single actionable step in a learning plan.
    Psychotype:
      type: string
      enum:
      - aggressive
      - confused
      - technically_inept
      - fraudster
      - neutral
      title: Psychotype
      description: Client psychotype for the simulation.
    QuizQuestion:
      properties:
        question:
          type: string
          minLength: 1
          title: Question
        options:
          items:
            type: string
          type: array
          minItems: 2
          title: Options
        correct_index:
          type: integer
          minimum: 0.0
          title: Correct Index
        explanation:
          type: string
          title: Explanation
          default: ''
      type: object
      required:
      - question
      - options
      - correct_index
      title: QuizQuestion
      description: A single question in a micro-quiz.
    QuizRequest:
      properties:
        scenario_id:
          type: string
          format: uuid
          title: Scenario Id
        question_count:
          type: integer
          title: Question Count
          default: 5
      type: object
      required:
      - scenario_id
      title: QuizRequest
    QuizResponse:
      properties:
        quiz:
          $ref: '#/components/schemas/MicroQuiz'
      type: object
      required:
      - quiz
      title: QuizResponse
...
    RespondRequest:
      properties:
        session_id:
          type: string
          format: uuid
          title: Session Id
      type: object
      required:
      - session_id
      title: RespondRequest
    RespondResponse:
      properties:
        client_message:
          type: string
          title: Client Message
      type: object
      required:
      - client_message
      title: RespondResponse
...
    StartDialogueRequest:
      properties:
        scenario_id:
          type: string
          format: uuid
          title: Scenario Id
      type: object
      required:
      - scenario_id
      title: StartDialogueRequest
    StartDialogueResponse:
      properties:
        greeting:
          type: string
          title: Greeting
        psychotype:
          $ref: '#/components/schemas/Psychotype'
      type: object
      required:
      - greeting
      - psychotype
      title: StartDialogueResponse
...
    SyncLMSRequest:
      properties:
        plan_id:
          type: string
          format: uuid
          title: Plan Id
      type: object
      required:
      - plan_id
      title: SyncLMSRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError